# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: The race cars
  contact:
    email: loi.tra@9thwonder.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/
  - url: https://virtserver.swaggerhub.com/annguyen9w/hapi-starter/1.0.0
    description: SwaggerHub API Auto Mocking
paths:
...
  /teams:
    get:
      tags:
        - Team
      summary: Get all team
      operationId: getTeams
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        # '400':
        #   $ref: '#/components/responses/badrequest'
    post:
      tags:
        - Team
      summary: Add a new team to the system
      operationId: addTeam
      requestBody:
        $ref: '#/components/requestBodies/Team'
      responses:
        '200':
          $ref: '#/components/responses/successwithid'
        '400':
          $ref: '#/components/responses/badrequest'
  /teams/{teamId}:
    get:
      tags:
        - Team
      summary: Find team by ID
      operationId: getTeamById
      parameters:
        - name: teamId
          in: path
          description: ID of team to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    put:
      tags:
        - Team
      summary: Update an existing team
      operationId: updateTeam
      parameters:
        - name: teamId
          in: path
          description: Team id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/Team'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Team
      summary: Delete a team
      operationId: deleteTeam
      parameters:
        - name: teamId
          in: path
          description: Team id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      example: 40e6215d-b5c6-4896-987c-f30f3678f608
    ClassRequestBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: LM GTE AM
    Class:
      allOf:
        - required:
            - id
          type: object
          properties:
            id:
              $ref: '#/components/schemas/UUID'
        - $ref: '#/components/schemas/ClassRequestBody'
    AddressRequestBody:
      required:
        - city
        - country
        - state
        - zipcode
      type: object
      properties:
        street:
          type: string
          example: 1797 Colonial Drive
        street2:
          type: string
          example: 411 Patterson Street
        city:
          type: string
          example: Houston
        state:
          type: string
          example: Texas
        zipcode:
          type: string
          example: '77002'
        country:
          type: string
          example: United States
    Address:
      allOf:
        - required:
            - id
          type: object
          properties:
            id:
              $ref: '#/components/schemas/UUID'
        - $ref: '#/components/schemas/AddressRequestBody'
...
    Driver:
      required:
        - firstName
        - id
        - lastName
        - nationality
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        firstName:
          type: string
          example: Nguyen
        lastName:
          type: string
          example: An
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        homeAddress:
          $ref: '#/components/schemas/Address'
        managementAddress:
          $ref: '#/components/schemas/Address'
...
    TeamRequestBody:
      required:
        - name
        - nationality
      type: object
      properties:
        name:
          type: string
          example: 7 TOYOTA GAZOO RACING
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        businessAddress:
          $ref: '#/components/schemas/UUID'
        drivers:
          type: array
          items:
            $ref: '#/components/schemas/UUID'
    Team:
      required:
        - cars
        - id
        - name
        - nationality
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: 7 TOYOTA GAZOO RACING
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        businessAddress:
          $ref: '#/components/schemas/Address'
    TeamResponse:
      allOf:
        - $ref: '#/components/schemas/Team'
        - required:
            - car
          type: object
          properties:
            cars:
              type: array
              items:
                $ref: '#/components/schemas/Car'
            drivers:
              type: array
              items:
                $ref: '#/components/schemas/Driver'
...
    Car:
      required:
        - class
        - make
        - model
        - team
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        make:
          type: string
          example: Ferrari
        model:
          type: string
          example: 488 GTE-LM
        class:
          $ref: '#/components/schemas/Class'
...
  requestBodies:
...
    Team:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TeamRequestBody'
      required: true
...
  responses:
    success:
      description: Successful operation
      content:
        text/plain:
          schema:
            type: string
            example: success
    successwithid:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              id: 
                $ref: '#/components/schemas/UUID'
              status:
                type: string
                default: success
            required:
              - id
              - status
    badrequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: 
                type: integer
                default: 400
              error:
                type: string
                default: Bad Request
              message:
                type: string
                default: Invalid request input
            required:
              - statusCode
              - error
              - message
    notfound:
      description: Item not found
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: 
                type: integer
                default: 404
              error:
                type: string
                default: Not Found
              message:
                type: string
                default: Not Found
            required:
              - statusCode
              - error
              - message